Moved developerEmbed to other folder
