+++ commands/invites.ts
@@ -1,8 +1,8 @@
 import { SlashCommandBuilder } from '@discordjs/builders';
-import { CommandInteraction, MessageEmbed } from 'discord.js';
+import { CommandInteraction } from 'discord.js';
 import { client } from '..';
 import { supabase } from '../database';
-import { createDeveloperEmbed } from '../utils/developerEmbed';
+import { createDeveloperEmbed } from '../components/developerEmbed';
 const paginationEmbed = require('../components/paginationEmbed.js')
 
 export const data = new SlashCommandBuilder()
@@ -57,19 +57,6 @@ export async function execute(interaction: CommandInteraction) {
         if (data!.length === 0) await interaction.editReply({
             content: 'You have no invites.'
         })
-
-        const { 
-            discord_id,
-            timezone,
-            discord, 
-            position, 
-            skills, 
-            desired_skills, 
-            goal,
-            available,
-            github,
-            twitter
-        } = data![0]
     
         const user = await client.users?.fetch(data![0].discord_id);
         senderDiscordIds.push(user.id);
